requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
waitress==3.0.2
